Add option to read nrf51 UICR memory, included in GUI
Adds the option to read the UICR Customer memory location.
Add that to the GUI.
Moves the GUI menus around to have "micro:bit" (for reading MicroPytho
and the Python code) and "nrf" (for nrf51 specific commands, like to
read all flash, or the UICR registers).

Fixed issue with pretty print to showing the correct start address
(it always started at 0x00).